refactor split app and server
