+++ app/build.gradle
@@ -1,0 +1,39 @@
+plugins {
+    id 'application'
+    id 'com.github.johnrengelman.shadow' version '6.0.0'
+    id 'com.adarshr.test-logger' version '2.1.0'
+}
+
+group 'dev.keva'
+version '1.0-SNAPSHOT'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation project(':server')
+    implementation project(':config')
+
+    implementation 'ch.qos.logback:logback-classic:1.2.7'
+    implementation 'ch.qos.logback:logback-core:1.2.7'
+
+    testImplementation 'redis.clients:jedis:3.7.0'
+}
+
+shadowJar {
+    minimize()
+}
+
+application {
+    mainClassName = 'dev.keva.app.Application'
+}
+
+test {
+    useJUnitPlatform()
+    testLogging {
+        outputs.upToDateWhen { true }
+        showStandardStreams = true
+    }
+    maxParallelForks = Runtime.runtime.availableProcessors()
+}
