+++ server/build.gradle
@@ -1,10 +1,8 @@
 plugins {
-    id 'application'
-    id 'com.github.johnrengelman.shadow' version '6.0.0'
     id 'com.adarshr.test-logger' version '2.1.0'
 }
 
-group 'com.jinyframework'
+group 'dev.keva'
 version '1.0-SNAPSHOT'
 
 repositories {
@@ -12,12 +10,10 @@ repositories {
 }
 
 dependencies {
-    implementation project(':config')
-    implementation project(':store')
     implementation project(':protocol')
+    implementation project(':store')
+    implementation project(':config')
 
-    implementation 'ch.qos.logback:logback-classic:1.2.7'
-    implementation 'ch.qos.logback:logback-core:1.2.7'
     implementation 'io.netty:netty-handler:4.1.70.Final'
     implementation 'io.netty:netty-buffer:4.1.70.Final'
     implementation 'io.netty:netty-transport-native-kqueue:4.1.70.Final:osx-x86_64'
@@ -25,9 +21,8 @@ dependencies {
 
     implementation 'com.google.guava:guava:31.0.1-jre'
     implementation 'org.reflections:reflections:0.10.2'
-    implementation 'org.apache.commons:commons-lang3:3.0'
+    implementation 'org.apache.commons:commons-lang3:3.12.0'
 
-    // Experimental
     implementation files('libs/keva-ioc-0.1.0-SNAPSHOT.jar')
 
     testImplementation 'redis.clients:jedis:3.7.0'
@@ -36,14 +31,6 @@ dependencies {
     testImplementation group: 'junit', name: 'junit', version: '4.12'
     // noinspection GradlePackageUpdate
     testRuntimeOnly('org.junit.vintage:junit-vintage-engine:5.7.2')
-}
-
-shadowJar {
-    minimize()
-}
-
-application {
-    mainClassName = 'dev.keva.server.Application'
 }
 
 test {
